--- Assets/Domain/Monster/MonsterAI/NavMesh.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dac1bd64dd086c64e869eb6dc93b0199
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
