--- Assets/Domain/Monster/MonsterAnimator.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 71e2f770ced29a343a4f61079d613629
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
