--- Assets/Domain/Monster/MonsterController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d6d780de6fa24ac428806be2feee7251
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
