+++ Assets/Domain/Scripts/NetworkScripts/OwnershipTransfer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a138ad723cf4ec74baedf92d1e9c883c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
