+++ Assets/Domain/Scripts/UI/TitleUI.cs
@@ -13,7 +13,6 @@ public class TitleUI : MonoBehaviourPunCallbacks
 
     private void Awake()
     {
-        //Screen.SetResolution(760, 480, true);
     }
     public void OnClickOnlineBtn()
     {
