+++ Assets/Editor/Monster.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c054dab0ee5f434f81856b790f3b083
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
