+++ Assets/Editor/Monster/MonsterAI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02ab3e3466379f14f8ed2b91a29c71e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
