+++ Assets/Editor/Monster/PlayerPreviousTransform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c00002e5a5a99048863ab814a2ee326
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
