+++ Assets/Editor/Monster/Waypoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3ee6f771d7ed464cbbb0d54e5acc4f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
