+++ Assets/Editor/Monster/view.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2609715d0ea61e74da35760f02a7c66e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
