+++ Assets/Editor/Monster/view/MonsterView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 185dc5f416a1bc845bbbd26dd77ade74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
