+++ Assets/PhotonTest-KKB.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7b18114acc18ed949a84942a37833bd9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
