+++ Assets/Resources/New Prefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c157f1b6fd6cb7d409e5a06db02cfb8c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
