+++ Assets/Resources/PlayerLatifa.prefab
@@ -1491,7 +1491,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1758405483690713097}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 1.375, z: -0.16666678}
+  m_LocalPosition: {x: 0, y: 1.3749998, z: -0.16666625}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -3147,8 +3147,8 @@ GameObject:
   - component: {fileID: 4079919205582287008}
   - component: {fileID: 4939969789357892556}
   - component: {fileID: 7008048694918821904}
-  - component: {fileID: 5988903934231497993}
   - component: {fileID: 2655993273688867196}
+  - component: {fileID: 7830221213572229074}
   - component: {fileID: 4739832367753883608}
   - component: {fileID: 369565989063905590}
   m_Layer: 8
@@ -3252,7 +3252,10 @@ MonoBehaviour:
   InventoryManager: {fileID: 2312498499568412951}
   Minimap: {fileID: 8870119579038678174}
   CurrentMap: {fileID: 0}
-  TextInteraction: {fileID: 1477754087404381693}
+  ItemInteraction: {fileID: 0}
+  LockInteraction: {fileID: 0}
+  LockView: {fileID: 0}
+  ItemView: {fileID: 0}
   InvestigateValue: 0
 --- !u!114 &7981337102398112960
 MonoBehaviour:
@@ -3337,6 +3340,8 @@ MonoBehaviour:
   interaction: 0
   inventory: 0
   minimap: 0
+  LookLock: 0
+  MoveLock: 0
   UILock: 0
   analogMovement: 0
   cursorLocked: 1
@@ -3440,42 +3445,6 @@ MonoBehaviour:
   m_DefaultActionMap: Player
   m_SplitScreenIndex: -1
   m_Camera: {fileID: 0}
---- !u!114 &5988903934231497993
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3817298006141453325}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9b8c4a61274f60b4ea5fb4299cfdbf14, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ShowLayerWeightsInspector: 1
-  ShowParameterInspector: 1
-  m_SynchronizeParameters:
-  - Type: 1
-    SynchronizeType: 0
-    Name: Speed
-  - Type: 4
-    SynchronizeType: 0
-    Name: Jump
-  - Type: 4
-    SynchronizeType: 0
-    Name: Grounded
-  - Type: 4
-    SynchronizeType: 0
-    Name: FreeFall
-  - Type: 1
-    SynchronizeType: 0
-    Name: MotionSpeed
-  - Type: 4
-    SynchronizeType: 0
-    Name: Investigate
-  m_SynchronizeLayers:
-  - SynchronizeType: 0
-    LayerIndex: 0
 --- !u!114 &2655993273688867196
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -3492,6 +3461,18 @@ MonoBehaviour:
   m_SynchronizeRotation: 1
   m_SynchronizeScale: 0
   m_UseLocal: 1
+--- !u!114 &7830221213572229074
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3817298006141453325}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a138ad723cf4ec74baedf92d1e9c883c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &4739832367753883608
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -3511,7 +3492,6 @@ MonoBehaviour:
   OwnershipTransfer: 1
   observableSearch: 2
   ObservedComponents:
-  - {fileID: 5988903934231497993}
   - {fileID: 2655993273688867196}
   sceneViewId: 0
   InstantiationId: 0
@@ -5177,7 +5157,9 @@ MonoBehaviour:
   ItemContent: {fileID: 4604799300649834788}
   InventoryItem: {fileID: 3424938949494706469, guid: af8c1eb66f077214090c9f0c667d3b8f,
     type: 3}
+  SelectedItem: {fileID: 0}
   InventoryItemControllers: []
+  ItemView: {fileID: 0}
 --- !u!1 &5991133848980984079
 GameObject:
   m_ObjectHideFlags: 0
@@ -7977,12 +7959,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: d435b4c65d9874542bee7c2ecd2ce030, type: 3}
---- !u!1 &1477754087404381693 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 5721627802026848136, guid: d435b4c65d9874542bee7c2ecd2ce030,
-    type: 3}
-  m_PrefabInstance: {fileID: 6621777616981559925}
-  m_PrefabAsset: {fileID: 0}
 --- !u!224 &1477754087404381694 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 5721627802026848139, guid: d435b4c65d9874542bee7c2ecd2ce030,
