+++ Assets/Resources/PlayerLatifaKWJ.prefab
@@ -3234,6 +3234,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 006bddaaf4777254696ff9c0a3da934c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  pv: {fileID: 4739832367753883608}
 --- !u!114 &6009440614073280171
 MonoBehaviour:
   m_ObjectHideFlags: 0
