+++ ProjectSettings/EditorBuildSettings.asset
@@ -14,15 +14,18 @@ EditorBuildSettings:
   - enabled: 0
     path: 
     guid: 00000000000000000000000000000000
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/testSceneKWJ.unity
     guid: c7f1076db8e39c0449bda875140b4cdf
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/testScene.unity
     guid: 67206306b634a1d45bb19d1374a1f57d
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/cutScene.unity
     guid: 46b46a276b7492a4f8f7047043dfa0a6
+  - enabled: 1
+    path: Assets/PhotonTest-KKB.unity
+    guid: 7b18114acc18ed949a84942a37833bd9
   m_configObjects:
     com.unity.input.settings: {fileID: 11400000, guid: 9e7be553448fa2546aea5752021cbcf7,
       type: 2}
